refactor: extract getAssistantName to shared identity-helpers
